+++ BIBIGOproject/BIBIGOproject/urls.py
@@ -19,11 +19,12 @@ from django.contrib import admin
 from django.urls import path, include
 from django.conf import settings
 from django.conf.urls.static import static
-
+from django.views.generic import RedirectView
 
 urlpatterns = [
     path("admin/", admin.site.urls),
     path("myapp/", include('myapp.urls')),  # 連接應用層級的 urls.py  
+    path("", RedirectView.as_view(url="/myapp/subpage/")),  # 首頁重定向
 ]
 
 # 這行只在開發模式 (DEBUG=True) 時有效，正式環境不會生效
