+++ BIBIGOproject/static/base.css
@@ -1,4 +1,3 @@
-/* 基本樣式 */
 * {
     margin: 0;
     padding: 0;
@@ -6,111 +5,108 @@
 }
 
 body {
-    font-family: Arial, sans-serif;
-    background-color: #f4f4f4;
-    color: #333;
+    font-family: "Roboto", sans-serif;
+    background-color: #fff;
+    color: #030303;
+    line-height: 1.5;
     display: flex;
     flex-direction: column;
     min-height: 100vh;
 }
 
-/* 頁首樣式 */
+
 .header {
-    background-color: #ffffff;
-    padding: 16px 0;
-    border-bottom: 1px solid #ddd;
-    position: relative;
+    background-color: #fff;
+    padding: 8px 16px;
+    border-bottom: 1px solid #e0e0e0;
+    position: sticky;
+    top: 0;
+    z-index: 1000;
 }
 
 .container {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
+    max-width: 1440px;
     margin: 0 auto;
-    padding: 0 50px;
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+    gap: 16px;
+    padding: 0 16px;
 }
 
 .logo {
-    position: absolute;
-    left: 20px;
-    z-index: 1;
+    flex-shrink: 0;
 }
 
 .logo-img {
     max-width: 100px;
+    height: auto;
 }
 
 .nav {
-    display: flex;
-    gap: 20px;
-    margin-left: 120px;
-    justify-content: flex-end;
-    position: relative;
-    z-index: 2;
+    flex-grow: 1;
+    margin-left: 24px;
 }
 
 .nav-links {
     list-style-type: none;
     display: flex;
-    gap: 30px;
+    gap: 24px;
 }
 
-.nav-links li {
-    display: inline-block;
-}
-
-.nav-links a {
+.nav-links li a {
     text-decoration: none;
-    color: #333;
-    font-size: 18px;
-    font-weight: bold;
-    transition: color 0.3s;
+    color: #030303;
+    font-size: 20px; 
+    font-weight: 700; 
+    padding: 8px 12px;
+    transition: background-color 0.2s;
 }
 
-.nav-links a:hover {
-    color: #007bff;
+.nav-links li a:hover {
+    background-color: #f1f1f1;
+    border-radius: 4px;
 }
 
-/* 登入註冊按鈕 */
+
 .auth-buttons {
-    position: absolute;
-    right: 20px;
+    flex-shrink: 0;
 }
 
 .btn-auth {
     background-color: #007bff;
     color: #fff;
     border: none;
-    padding: 10px 20px;
+    padding: 12px 24px;
     font-size: 16px;
-    border-radius: 4px;
+    border-radius: 24px;
     cursor: pointer;
-    transition: background-color 0.3s;
     text-decoration: none;
     display: inline-block;
+    transition: background-color 0.2s;
 }
 
 .btn-auth:hover {
     background-color: #0056b3;
-    text-decoration: none;
 }
 
-/* 頁尾樣式 */
+
+.main-content {
+    flex: 1 0 auto;
+    padding: 16px 0;
+}
+
+
 .footer {
-    background-color: #ffffff;
+    background-color: #f8f8f8;
     text-align: center;
-    padding: 20px;
-    border-top: 1px solid #ddd;
-    width: 100%;
-    margin-top: auto;
+    padding: 16px;
+    border-top: 1px solid #e0e0e0;
+    color: #606060;
+    font-size: 14px;
 }
 
-.footer p {
-    color: #777;
-    margin: 0;
-}
 
-/* 人工客服浮動樣式 */
 .customer-service-floating {
     position: fixed;
     bottom: 20px;
@@ -119,8 +115,8 @@ body {
 }
 
 .customer-service-icon {
-    width: 100px;
-    height: 100px;
+    width: 80px;
+    height: 80px;
     cursor: pointer;
     transition: transform 0.2s;
 }
@@ -131,13 +127,13 @@ body {
 
 .customer-service-content {
     position: fixed;
-    bottom: 130px;
+    bottom: 90px;
     right: 20px;
-    width: 350px;
+    width: 320px;
     background: #fff;
-    border-radius: 10px;
-    box-shadow: 0 2px 5px rgba(0, 0, 0, 0.1);
-    padding: 15px;
+    border-radius: 8px;
+    box-shadow: 0 2px 10px rgba(0, 0, 0, 0.1);
+    padding: 16px;
 }
 
 .customer-service {
@@ -146,9 +142,9 @@ body {
 }
 
 .service-title {
-    font-size: 18px;
-    font-weight: bold;
-    margin-bottom: 10px;
+    font-size: 16px;
+    font-weight: 500;
+    margin-bottom: 12px;
     text-align: center;
 }
 
@@ -159,97 +155,100 @@ body {
 
 .chat-input textarea {
     width: 100%;
-    height: 100px;
-    padding: 10px;
-    border: 1px solid #ccc;
-    border-radius: 5px;
+    height: 80px;
+    padding: 8px;
+    border: 1px solid #dadce0;
+    border-radius: 4px;
     resize: none;
     font-size: 14px;
-    margin-bottom: 10px;
+    margin-bottom: 8px;
 }
 
 .chat-input button {
-    background-color: #007bff;
+    background-color: #dc3545;
     color: #fff;
     border: none;
-    padding: 10px;
-    border-radius: 5px;
+    padding: 8px;
+    border-radius: 4px;
     cursor: pointer;
-    font-size: 16px;
-    transition: background-color 0.3s;
+    font-size: 14px;
+    transition: background-color 0.2s;
 }
 
 .chat-input button:hover {
-    background-color: #0056b3;
+    background-color: #c82333;
 }
 
-/* 響應式設計 */
-@media (max-width: 768px) {
-    .container {
-        flex-direction: column;
-        align-items: flex-start;
-    }
 
-    .nav {
-        margin-left: 0;
-        flex-direction: column;
-        gap: 15px;
-        margin-top: 15px;
-    }
-
-    .auth-buttons {
-        margin-top: 20px;
-        align-items: flex-start;
-        position: static;
-    }
-
-    .btn-auth {
-        width: 100%;
-        padding: 12px 20px;
-        font-size: 18px;
-    }
-
-    .customer-service-icon {
-        width: 50px;
-        height: 50px;
-    }
-
-    .customer-service-content {
-        bottom: 70px;
-        width: 300px;
-    }
-}
-
-/* 回到頂部按鈕 */
 #back-to-top {
-    display: none; /* 初始隱藏 */
+    display: none;
     position: fixed;
     bottom: 20px;
-    right: 140px; /* 避免與客服按鈕重疊 */
+    right: 100px;
     width: 50px;
     height: 50px;
     background-color: #007bff;
-    color: white;
+    color: #fff;
     border: none;
     border-radius: 50%;
     text-align: center;
-    line-height: 50px;
+    line-height: 40px;
     cursor: pointer;
-    box-shadow: 0 2px 5px rgba(0, 0, 0, 0.3);
+    box-shadow: 0 2px 5px rgba(0, 0, 0, 0.2);
     z-index: 1000;
-    transition: background-color 0.3s;
+    transition: background-color 0.2s;
 }
 
 #back-to-top:hover {
     background-color: #0056b3;
 }
 
-/* 響應式設計 - 回到頂部按鈕 */
-@media (max-width: 768px) {
+/* 響應式設計 */
+@media (max-width: 960px) {
+    .container {
+        flex-wrap: wrap;
+        gap: 8px;
+    }
+    .nav {
+        width: 100%;
+        margin-left: 0;
+        padding: 8px 0;
+    }
+    .nav-links {
+        flex-direction: column;
+        gap: 16px;
+        text-align: center;
+    }
+    .auth-buttons {
+        width: 100%;
+        text-align: center;
+        padding: 8px 0;
+    }
+    .btn-auth {
+        width: auto;
+    }
+}
+
+@media (max-width: 600px) {
+    .logo-img {
+        max-width: 80px;
+    }
+    .nav-links li a {
+        font-size: 14px;
+        padding: 6px 10px;
+    }
+    .customer-service-icon {
+        width: 50px;
+        height: 50px;
+    }
+    .customer-service-content {
+        bottom: 70px;
+        width: 280px;
+    }
     #back-to-top {
-        right: 70px;
-        width: 40px;
-        height: 40px;
-        line-height: 40px;
+        right: 80px;
+        width: 36px;
+        height: 36px;
+        line-height: 36px;
     }
 }
