+++ BIBIGOproject/static/care.css
@@ -1,153 +1,153 @@
-/* care.css */
 .subpage-container {
-    max-width: 1900px;
+    max-width: 1440px;
     width: 100%;
     margin: 0 auto;
+    padding: 0 16px;
 }
 
 .subpage-content-container {
     width: 100%;
-    max-width: 1900px;
-    min-width: 960px;
     margin: 0 auto;
 }
 
 .subpage-row {
     display: flex;
     flex-wrap: wrap;
-    justify-content: center;
-    width: 100%;
-    margin-top: 10px;
+    gap: 16px;
 }
 
+
 .subpage-sidebar {
-    flex: 1 1 200px;
-    max-width: 20%;
-    padding: 20px;
-    background-color: transparent;
+    flex: 0 0 240px;
+    padding: 16px 0;
     font-size: 20px;
 }
 
 .subpage-list-unstyled {
     list-style-type: none;
     padding-left: 0;
-    margin-top: 5px;
+    margin: 0;
 }
 
 .subpage-category {
+    padding: 8px 16px;
     cursor: pointer;
-    margin-bottom: 10px;
 }
 
 .subpage-category a {
     text-decoration: none;
-    color: #333;
+    color: #030303;
+    font-size: 20px;
+    font-weight: 500;
 }
 
 .subpage-category:hover {
-    background-color: #f8f9fa;
+    background-color: #f1f1f1;
 }
+
 
 .subpage-content {
-    flex: 2 1 600px;
-    max-width: 100%;
-    padding: 20px;
-    font-size: 20px;
+    flex: 1 1 0;
+    padding: 16px 0;
+    font-size: 16px;
 }
 
-/* 分類器樣式 */
+
 .care-filter {
-    margin: 20px 0;
-    padding: 10px;
-    background: #f8f9fa;
-    border-radius: 5px;
+    margin: 16px 0;
+    padding: 8px;
+    background: #fff;
+    border-radius: 8px;
 }
 
 .filter-row {
-    margin-bottom: 15px;
+    margin-bottom: 16px;
 }
 
 .filter-row h3 {
     font-size: 16px;
-    margin-bottom: 10px;
-    color: #333;
+    margin-bottom: 8px;
+    color: #030303;
 }
 
 .filter-options {
     display: flex;
     flex-wrap: wrap;
-    gap: 10px;
+    gap: 8px;
 }
 
 .filter-btn {
-    padding: 8px 15px;
-    background: #ddd;
-    border-radius: 20px;
+    padding: 6px 16px; 
+    background: #e0e0e0;
+    border-radius: 16px;
     text-decoration: none;
-    color: #333;
-    font-size: 14px;
-    transition: background 0.3s;
+    color: #030303;
+    font-size: 16px; 
+    transition: background-color 0.2s;
 }
 
 .filter-btn:hover {
-    background: #ccc;
+    background: #d0d0d0;
 }
 
 .filter-btn.active {
     background: #ff5722;
-    color: white;
+    color: #fff;
 }
 
-/* 推薦商品區域 */
+
 .recommended-products {
-    margin: 20px 0;
+    margin: 16px 0;
 }
 
 .recommended-products h3 {
     font-size: 18px;
-    font-weight: bold;
-    margin-bottom: 10px;
+    font-weight: 500;
+    margin-bottom: 8px;
 }
 
 .recommended-container {
-    position: relative;
-    display: flex;
-    align-items: center;
+    width: 100%;
 }
 
 .recommended-list {
     display: flex;
-    gap: 10px; /* 減少間距以容納更多商品 */
-    width: 100%;
-    overflow: hidden;
+    align-items: center;
+    gap: 16px;
+    overflow-x: auto;
+    -webkit-overflow-scrolling: touch;
+    scroll-behavior: smooth;
 }
 
 .recommended-item {
     flex: 0 0 auto;
-    width: 80px; /* 調整寬度以適應 10 個商品 */
-    text-align: center;
+    width: 140px;
+    padding: 8px;
+    background-color: #fff;
+    border-radius: 8px;
 }
 
 .recommended-link {
     text-decoration: none;
-    color: #333;
+    color: #030303;
     display: block;
 }
 
 .recommended-item img {
     width: 100%;
-    height: 80px; /* 調整圖片高度 */
+    height: auto;
+    aspect-ratio: 16 / 9;
+    border-radius: 8px;
     object-fit: cover;
-    border-radius: 5px;
 }
 
 .recommended-info {
-    margin-top: 5px;
+    margin: 8px 0;
 }
 
 .recommended-name {
     display: block;
-    font-size: 12px; /* 調整字體大小 */
+    font-size: 14px;
     white-space: nowrap;
     overflow: hidden;
     text-overflow: ellipsis;
@@ -155,19 +155,23 @@
 
 .recommended-price {
     display: block;
-    font-size: 12px;
+    font-size: 14px;
     color: #dc3545;
-    margin-top: 2px;
+    font-weight: 500;
 }
 
 .nav-btn {
+    flex: 0 0 auto;
     background: #f1f1f1;
     border: none;
     font-size: 18px;
-    padding: 10px;
+    padding: 8px 12px;
     cursor: pointer;
     border-radius: 5px;
-    transition: background 0.3s;
+    height: 40px;
+    display: flex;
+    align-items: center;
+    justify-content: center;
 }
 
 .nav-btn:hover {
@@ -175,42 +179,29 @@
 }
 
 .nav-btn:disabled {
-    background: #f1f1f1;
-    cursor: not-allowed;
     opacity: 0.5;
+    cursor: not-allowed;
 }
 
-.prev-btn {
-    margin-right: 10px;
-}
 
-.next-btn {
-    margin-left: 10px;
-}
-
-/* 商品列表樣式 */
 .subpage-products {
     display: flex;
     flex-direction: column;
-    gap: 0;
+    gap: 16px;
 }
 
 .followed-product-item {
     display: flex;
     align-items: center;
-    gap: 15px;
-    padding: 15px;
+    gap: 16px;
+    padding: 16px; 
     background: #fff;
-    border-radius: 0;
-    box-shadow: 0 1px 3px rgba(0, 0, 0, 0.1);
-    width: 100%;
-    box-sizing: border-box;
-    margin: 0;
-    border-bottom: 1px solid #ddd;
+    border-radius: 8px;
+    transition: background-color 0.2s;
 }
 
-.followed-product-item:last-child {
-    border-bottom: none;
+.followed-product-item:hover {
+    background-color: #f5f5f5;
 }
 
 .rank {
@@ -218,51 +209,54 @@
     color: #ff5722;
     min-width: 20px;
     text-align: center;
+    font-size: 16px; 
 }
 
 .store-logo {
-    width: 40px;
-    height: 40px;
+    width: 48px; 
+    height: 48px; 
     object-fit: contain;
 }
 
 .product-link {
     display: flex;
     align-items: center;
-    gap: 15px;
+    gap: 16px;
     text-decoration: none;
-    color: #333;
+    color: #030303;
     flex-grow: 1;
 }
 
 .product-name {
     text-align: left;
-    font-size: 16px;
-    color: #333;
+    font-size: 16px; 
+    color: #030303;
 }
 
 .right-aligned {
     display: flex;
     align-items: center;
-    gap: 15px;
+    gap: 16px;
     margin-left: auto;
 }
 
 .product-price {
-    min-width: 100px;
+    min-width: 80px;
     text-align: left;
-    font-weight: bold;
+    font-weight: 500;
     color: #dc3545;
+    font-size: 16px; 
 }
 
 .buy-btn {
     background: #ff9800;
-    color: white;
+    color: #fff;
     border: none;
-    padding: 5px 10px;
+    padding: 6px 12px; 
     border-radius: 5px;
     cursor: pointer;
     text-decoration: none;
+    font-size: 16px; 
 }
 
 .buy-btn:hover {
@@ -291,40 +285,56 @@
     background-size: contain;
 }
 
-@media (max-width: 992px) {
-    .subpage-sidebar, .subpage-content {
-        max-width: 100%;
-        flex: 1 1 100%;
+/* 響應式設計*/
+@media (max-width: 1312px) {
+    .subpage-sidebar {
+        flex: 0 0 72px;
+        text-align: center;
     }
+    .subpage-category {
+        padding: 16px 0;
+    }
+}
 
+@media (max-width: 960px) {
+    .subpage-sidebar {
+        display: none;
+    }
+    .subpage-content {
+        width: 100%;
+    }
     .followed-product-item {
         flex-wrap: wrap;
-        gap: 10px;
+        gap: 8px;
     }
-
     .right-aligned {
         margin-left: 0;
         justify-content: flex-start;
     }
+}
 
-    .care-filter {
-        padding: 10px;
-    }
-
+@media (max-width: 600px) {
     .filter-btn {
-        padding: 6px 12px;
+        padding: 4px 12px; 
+        font-size: 14px; 
+    }
+    .recommended-item {
+        width: 120px;
+    }
+    .recommended-name, .recommended-price {
         font-size: 12px;
     }
-
-    .recommended-item {
-        width: 60px; /* 小螢幕下進一步縮小 */
+    .followed-product-item {
+        padding: 12px; 
     }
-
-    .recommended-item img {
-        height: 60px;
+    .store-logo {
+        width: 40px; 
+        height: 40px;
     }
-
-    .recommended-name, .recommended-price {
-        font-size: 10px;
+    .product-name, .product-price, .buy-btn {
+        font-size: 14px; 
+    }
+    .buy-btn {
+        padding: 5px 10px; 
     }
 }
