+++ BIBIGOproject/static/forget_password.css
@@ -1,5 +1,3 @@
-/* forget_password.css */
-/* 忘記密碼容器 */
 .forget-password-container {
     max-width: 400px;
     width: 100%;
@@ -11,21 +9,21 @@
     text-align: center;
 }
 
-/* 標題 */
+
 .forget-password-container h1 {
     font-size: 24px;
     color: #333;
     margin-bottom: 20px;
 }
 
-/* 表單樣式 */
+
 #forget-password-form {
     display: flex;
     flex-direction: column;
     gap: 15px;
 }
 
-/* 輸入框 */
+
 #forget-password-form input {
     width: 100%;
     padding: 10px;
@@ -40,7 +38,7 @@
     border-color: #007bff;
 }
 
-/* 按鈕 */
+
 #forget-password-form button {
     padding: 10px;
     background-color: #007bff;
@@ -56,7 +54,7 @@
     background-color: #0056b3;
 }
 
-/* 訊息 */
+
 #message {
     margin-top: 15px;
     font-size: 14px;
@@ -70,7 +68,7 @@
     color: #dc3545;
 }
 
-/* 返回連結 */
+
 .back-link {
     display: inline-block;
     margin-top: 20px;
