+++ BIBIGOproject/static/login.css
@@ -89,39 +89,6 @@ hr {
     color: white;
 }
 
-.google {
-    background-color: #db4437;
-    color: white;
-    border: none;
-    padding: 10px 20px;
-    margin: 10px;
-    border-radius: 5px;
-    cursor: pointer;
-    font-size: 16px;
-}
-
-.facebook {
-    background-color: #3b5998;
-    color: white;
-    border: none;
-    padding: 10px 20px;
-    margin: 10px;
-    border-radius: 5px;
-    cursor: pointer;
-    font-size: 16px;
-}
-
-.google:hover {
-    background-color: #c23321;
-}
-
-.facebook:hover {
-    background-color: #2d4373;
-}
-
-.mac {
-    background-color: #333;
-}
 
 @media (max-width: 480px) {
     .container {
