+++ BIBIGOproject/static/register.css
@@ -1,5 +1,3 @@
-/* register.css */
-
 body {
     font-family: 'Arial', sans-serif;
     display: flex;
