+++ BIBIGOproject/templates/base.html
@@ -2,9 +2,9 @@
 <html lang="zh-Hant">
 <head>
     {% load static %}
-
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <link href="https://fonts.googleapis.com/css2?family=Roboto:wght@400;500&display=swap" rel="stylesheet">
     <link rel="stylesheet" href="{% static 'base.css' %}">
     {% block extra_styles %}
     <link rel="stylesheet" href="{% static 'subpage.css' %}">
@@ -42,8 +42,7 @@
 
     <!-- 主要內容區 -->
     <main class="main-content">
-        {% block content %}
-        {% endblock %}
+        {% block content %}{% endblock %}
     </main>
 
     <!-- 頁尾 -->
