+++ BIBIGOproject/templates/login.html
@@ -30,11 +30,6 @@
         <div class="forgot-password">
             <a href="{% url 'forget_password' %}">忘記密碼</a>
         </div>
-        <hr>
-        <div class="social-login">
-            <a href="/" id="google-login" class="google">Google 登入</a>
-            <a href="/" id="facebook-login" class="facebook">Facebook 登入</a>
-        </div>
     </div>
 </body>
 </html>
